bug 10670

+++ org.eclipse.jdt.debug/model/org/eclipse/jdt/internal/debug/core/model/JDIDebugTarget.java
@@ -565,7 +565,7 @@ public class JDIDebugTarget extends JDIDebugElement implements IJavaDebugTarget,
 	 * @return whether this debug target supports popping stack frames.
 	 */
 	public boolean canPopFrames() {
-		if (isAvailable() && JDIDebugPlugin.getJDIVersion() >= 1.4) {
+		if (isAvailable() && JDIDebugPlugin.getJDIVersion() >= (float)1.4) {
 			return getVM().canPopFrames();
 		}
 		return false;
