+++ org.eclipse.jdt.debug/model/org/eclipse/jdt/internal/debug/core/model/JDIStackFrame.java
@@ -850,7 +850,7 @@ public class JDIStackFrame extends JDIDebugElement implements IJavaStackFrame {
 	 * @see IJavaStackFrame#isObsolete()
 	 */
 	public boolean isObsolete() throws DebugException {
-		if (JDIDebugPlugin.getJDIVersion() < 1.4 || !((JDIDebugTarget)getDebugTarget()).hasHCROccurred()) {
+		if (JDIDebugPlugin.getJDIVersion() < (float)1.4 || !((JDIDebugTarget)getDebugTarget()).hasHCROccurred()) {
 			// If no hot code replace has occurred, this frame
 			// cannot be obsolete.
 			return false;
